chore: rename modules for better meaning and searching
