+++ packages/next/src/server/patch-error-inspect.ts
@@ -4,7 +4,7 @@ import * as url from 'url'
 import type * as util from 'util'
 import { SourceMapConsumer as SyncSourceMapConsumer } from 'next/dist/compiled/source-map'
 import type { StackFrame } from 'next/dist/compiled/stacktrace-parser'
-import { parseStack } from '../client/components/react-dev-overlay/server/middleware'
+import { parseStack } from '../client/components/react-dev-overlay/server/middleware-webpack'
 import { getOriginalCodeFrame } from '../client/components/react-dev-overlay/server/shared'
 import { workUnitAsyncStorage } from './app-render/work-unit-async-storage.external'
 import { dim } from '../lib/picocolors'
