+++ packages/next/src/server/web/sandbox/context.ts
@@ -29,7 +29,7 @@ interface ModuleContext {
   warnedEvals: Set<string>
 }
 
-let getServerError: typeof import('../../../client/components/react-dev-overlay/server/middleware').getServerError
+let getServerError: typeof import('../../../client/components/react-dev-overlay/server/middleware-webpack').getServerError
 let decorateServerError: typeof import('../../../shared/lib/error-source').decorateServerError
 
 if (process.env.NODE_ENV === 'development') {
