+++ packages/next/src/server/web/sandbox/sandbox.ts
@@ -44,7 +44,7 @@ type RunnerFn = (params: RunnerFnParams) => Promise<FetchEventResult>
 function withTaggedErrors(fn: RunnerFn): RunnerFn {
   if (process.env.NODE_ENV === 'development') {
     const { getServerError } =
-      require('../../../client/components/react-dev-overlay/server/middleware') as typeof import('../../../client/components/react-dev-overlay/server/middleware')
+      require('../../../client/components/react-dev-overlay/server/middleware') as typeof import('../../../client/components/react-dev-overlay/server/middleware-webpack')
 
     return (params) =>
       fn(params)
